got docker image to work for the first time
